+++ tests/test_driver.py
@@ -18,92 +18,6 @@ logger = logging.getLogger(__name__)
 water_filepath = Path("tests/molecules/water.xyz").absolute()
 
 
-class UnrestrictedDriver(NbedDriver):
-    """Force use of unrestricted SCF for driver."""
-
-    def __init__(
-        self,
-        geometry: str,
-        n_active_atoms: int,
-        basis: str,
-        xc_functional: str,
-        projector: str,
-        localization: Optional[str] = "spade",
-        convergence: Optional[float] = 1e-6,
-        charge: Optional[int] = 0,
-        spin: Optional[int] = 0,
-        mu_level_shift: Optional[float] = 1e6,
-        run_ccsd_emb: Optional[bool] = False,
-        run_fci_emb: Optional[bool] = False,
-        run_virtual_localization: Optional[bool] = False,
-        run_dft_in_dft: Optional[bool] = False,
-        max_ram_memory: Optional[int] = 4000,
-        pyscf_print_level: int = 1,
-        savefile: Optional[Path] = None,
-        unit: Optional[str] = "angstrom",
-        occupied_threshold: Optional[float] = 0.95,
-        virtual_threshold: Optional[float] = 0.95,
-        init_huzinaga_rhf_with_mu: bool = False,
-        max_hf_cycles: int = 50,
-        max_dft_cycles: int = 50,
-    ):
-        """Initialise class."""
-        logger.debug("Initialising driver.")
-        config_valid = True
-        if projector not in ["mu", "huzinaga", "both"]:
-            logger.error(
-                "Invalid projector %s selected. Choose from 'mu' or 'huzinzaga'.",
-                projector,
-            )
-            config_valid = False
-
-        if localization not in ["spade", "ibo", "boys", "pipek-mezey"]:
-            logger.error(
-                "Invalid localization method %s. Choose from 'ibo','boys','pipek-mezey' or 'spade'.",
-                localization,
-            )
-            config_valid = False
-
-        if not config_valid:
-            logger.error("Invalid config.")
-            raise NbedConfigError("Invalid config.")
-
-        self.geometry = geometry
-        self.n_active_atoms = n_active_atoms
-        self.basis = basis.lower()
-        self.xc_functional = xc_functional.lower()
-        self.projector = projector.lower()
-        self.localization = localization.lower()
-        self.convergence = convergence
-        self.charge = charge
-        self.spin = spin
-        self.mu_level_shift = mu_level_shift
-        self.run_ccsd_emb = run_ccsd_emb
-        self.run_fci_emb = run_fci_emb
-        self.run_virtual_localization = run_virtual_localization
-        self.run_dft_in_dft = run_dft_in_dft
-        self.max_ram_memory = max_ram_memory
-        self.pyscf_print_level = pyscf_print_level
-        self.savefile = savefile
-        self.unit = unit
-        self.occupied_threshold = occupied_threshold
-        self.virtual_threshold = virtual_threshold
-        self.max_hf_cycles = max_hf_cycles
-        self.max_dft_cycles = max_dft_cycles
-
-        self._check_active_atoms()
-        self.localized_system = None
-        self.two_e_cross = None
-        self._dft_potential = None
-
-        self._restricted_scf = False
-
-        self.embed(
-            init_huzinaga_rhf_with_mu=init_huzinaga_rhf_with_mu
-        )  # TODO uncomment.
-        logger.debug("Driver initialisation complete.")
-
-
 def test_incorrect_geometry_path() -> None:
     """test to make sure that FileNotFoundError is thrown if invalid path to xyz geometry file is given"""
 
@@ -117,24 +31,13 @@ def test_incorrect_geometry_path() -> None:
         "projector": "mu",
         "localization": "spade",
         "convergence": 1e-6,
-        "savefile": None,
         "run_ccsd_emb": True,
         "run_fci_emb": True,
     }
 
     with pytest.raises(RuntimeError, match="Unsupported atom symbol .*"):
         # match will match with any printed error message
-        NbedDriver(
-            geometry=args["geometry"],
-            n_active_atoms=args["n_active_atoms"],
-            basis=args["basis"],
-            xc_functional=args["xc_functional"],
-            projector=args["projector"],
-            localization=args["localization"],
-            convergence=args["convergence"],
-            run_ccsd_emb=args["run_ccsd_emb"],
-            run_fci_emb=args["run_fci_emb"],
-        )
+        NbedDriver(**args)
 
 
 def test_driver_standard_xyz_file_input() -> None:
@@ -148,24 +51,13 @@ def test_driver_standard_xyz_file_input() -> None:
         "projector": "mu",
         "localization": "spade",
         "convergence": 1e-6,
-        "savefile": None,
-        "run_ccsd_emb": True,
-        "run_fci_emb": True,
+        "run_ccsd_emb": False,
+        "run_fci_emb": False,
     }
 
-    driver = NbedDriver(
-        geometry=args["geometry"],
-        n_active_atoms=args["n_active_atoms"],
-        basis=args["basis"],
-        xc_functional=args["xc_functional"],
-        projector=args["projector"],
-        localization=args["localization"],
-        convergence=args["convergence"],
-        run_ccsd_emb=args["run_ccsd_emb"],
-        run_fci_emb=args["run_fci_emb"],
-    )
+    driver = NbedDriver(**args)
     assert isinstance(driver.embedded_scf, StreamObject)
-    assert isinstance(driver.classical_energy, number)
+    assert isclose(driver.classical_energy, -14.229079481431608)
 
 
 def test_driver_standard_xyz_string_input() -> None:
@@ -175,33 +67,22 @@ def test_driver_standard_xyz_string_input() -> None:
     )
     args = {
         "geometry": water_xyz_raw,
-        "n_active_atoms": 1,
+        "n_active_atoms": 2,
         "basis": "STO-3G",
         "xc_functional": "b3lyp",
         "projector": "mu",
         "localization": "spade",
         "convergence": 1e-6,
-        "savefile": None,
         "run_ccsd_emb": False,
         "run_fci_emb": False,
     }
 
-    driver = NbedDriver(
-        geometry=args["geometry"],
-        n_active_atoms=args["n_active_atoms"],
-        basis=args["basis"],
-        xc_functional=args["xc_functional"],
-        projector=args["projector"],
-        localization=args["localization"],
-        convergence=args["convergence"],
-        run_ccsd_emb=args["run_ccsd_emb"],
-        run_fci_emb=args["run_fci_emb"],
-    )
+    driver = NbedDriver(**args)
     assert isinstance(driver.embedded_scf, StreamObject)
-    assert isinstance(driver.classical_energy, number)
+    assert isclose(driver.classical_energy, -3.5867934952241356)
 
 
-def test_n_active_atoms_valid() -> None:
+def test_n_active_atoms_validation() -> None:
     """test to check driver works... path to xyz file given"""
 
     args = {
@@ -211,37 +92,16 @@ def test_n_active_atoms_valid() -> None:
         "projector": "mu",
         "localization": "spade",
         "convergence": 1e-6,
-        "savefile": None,
         "run_ccsd_emb": True,
         "run_fci_emb": True,
     }
     error_msg = "Invalid number of active atoms. Choose a number from 1 to 2."
 
     with pytest.raises(NbedConfigError, match=error_msg):
-        NbedDriver(
-            geometry=args["geometry"],
-            n_active_atoms=0,
-            basis=args["basis"],
-            xc_functional=args["xc_functional"],
-            projector=args["projector"],
-            localization=args["localization"],
-            convergence=args["convergence"],
-            run_ccsd_emb=args["run_ccsd_emb"],
-            run_fci_emb=args["run_fci_emb"],
-        )
+        NbedDriver(n_active_atoms=0, **args)
 
     with pytest.raises(NbedConfigError, match=error_msg):
-        NbedDriver(
-            geometry=args["geometry"],
-            n_active_atoms=3,
-            basis=args["basis"],
-            xc_functional=args["xc_functional"],
-            projector=args["projector"],
-            localization=args["localization"],
-            convergence=args["convergence"],
-            run_ccsd_emb=args["run_ccsd_emb"],
-            run_fci_emb=args["run_fci_emb"],
-        )
+        NbedDriver(n_active_atoms=3, **args)
 
 
 def test_subsystem_dft() -> None:
@@ -254,22 +114,11 @@ def test_subsystem_dft() -> None:
         "projector": "mu",
         "localization": "spade",
         "convergence": 1e-6,
-        "savefile": None,
         "run_ccsd_emb": False,
         "run_fci_emb": False,
     }
 
-    driver = NbedDriver(
-        geometry=args["geometry"],
-        n_active_atoms=args["n_active_atoms"],
-        basis=args["basis"],
-        xc_functional=args["xc_functional"],
-        projector=args["projector"],
-        localization=args["localization"],
-        convergence=args["convergence"],
-        run_ccsd_emb=args["run_ccsd_emb"],
-        run_fci_emb=args["run_fci_emb"],
-    )
+    driver = NbedDriver(**args)
 
     energy_DFT_components = (
         driver.e_act
@@ -291,7 +140,6 @@ def test_subsystem_dft() -> None:
 #         "projector": "mu",
 #         "localization": "spade",
 #         "convergence": 1e-6,
-#         "savefile": None,
 #         "run_ccsd_emb": True,
 #         "run_fci_emb": True,
 #     }
@@ -304,7 +152,6 @@ def test_subsystem_dft() -> None:
 #         projector=args["projector"],
 #         localization=args["localization"],
 #         convergence=args["convergence"],
-#         savefile=args["savefile"],
 #         run_ccsd_emb=args["run_ccsd_emb"],
 #         run_fci_emb=args["run_fci_emb"],
 #     )
@@ -317,7 +164,6 @@ def test_subsystem_dft() -> None:
 #         projector=args["projector"],
 #         localization=args["localization"],
 #         convergence=args["convergence"],
-#         savefile=args["savefile"],
 #         run_ccsd_emb=args["run_ccsd_emb"],
 #         run_fci_emb=args["run_fci_emb"],
 #     )